+++ app/a4988/a4988.hpp
@@ -46,7 +46,9 @@ namespace a4988 {
 
         void gpio_init(this A4988 const& self) noexcept;
         void gpio_deinit(this A4988 const& self) noexcept;
-        void gpio_write_pin(this A4988 const& self, std::int16_t const pin, bool const state) noexcept;
+        void gpio_write_pin(this A4988 const& self,
+                            std::int32_t const pin,
+                            bool const state) noexcept;
     };
 
 }; // namespace a4988
